MB-67520: Fix many narrowing warnings in ep-engine
Various changes to address narrowing warnings in ep-engine

Warnings still remain after this commit.

Reviewed-on: https://review.couchbase.org/c/kv_engine/+/230758
Tested-by: Build Bot <build@couchbase.com>

+++ engines/ep/src/defragmenter.cc
@@ -216,12 +216,15 @@ DefragmenterTask::calculateSleepTimeAndRunState(
             true};
 }
 
-size_t DefragmenterTask::getAgeThreshold() const {
-    return engine->getConfiguration().getDefragmenterAgeThreshold();
+uint8_t DefragmenterTask::getAgeThreshold() const {
+    return gsl::narrow_cast<uint8_t>(
+            engine->getConfiguration().getDefragmenterAgeThreshold());
 }
 
-size_t DefragmenterTask::getStoredValueAgeThreshold() const {
-    return engine->getConfiguration().getDefragmenterStoredValueAgeThreshold();
+uint8_t DefragmenterTask::getStoredValueAgeThreshold() const {
+    return gsl::narrow_cast<uint8_t>(
+            engine->getConfiguration()
+                    .getDefragmenterStoredValueAgeThreshold());
 }
 
 void DefragmenterTask::updateStats(DefragmentVisitor& visitor) {
